refresh_token and all associated methods are moved to `grizzly.auth`
prepare for adding support for other clients than `RestApiUser`.
