+++ tests/__init__.py
@@ -1,0 +1,2 @@
+from gevent import monkey
+monkey.patch_all()
