+++ tests/unit/test_grizzly/auth/test___init__.py
@@ -6,15 +6,12 @@ import pytest
 from pytest_mock import MockerFixture
 
 from grizzly.auth import refresh_token, RefreshToken, AuthMethod, GrizzlyHttpAuthClient
-from grizzly.context import GrizzlyContextScenario
 from grizzly.users import RestApiUser
 from grizzly.tasks import RequestTask
 from grizzly.types import RequestMethod
 from grizzly.utils import safe_del
 
 from tests.fixtures import GrizzlyFixture
-
-RestApiScenarioFixture = Tuple[RestApiUser, GrizzlyContextScenario]
 
 
 class DummyAuth(RefreshToken):
@@ -23,24 +20,14 @@ class DummyAuth(RefreshToken):
         return 'dummy'
 
 
-@pytest.fixture
-def restapi_user(grizzly_fixture: GrizzlyFixture) -> RestApiScenarioFixture:
-    scenario = GrizzlyContextScenario(1)
-    scenario.name = scenario.description = 'TestScenario'
-    scenario.context['host'] = 'test'
-    scenario.user.class_name = 'RestApiUser'
-
-    _, user, _ = grizzly_fixture('http://example.net', RestApiUser)
-
-    request = grizzly_fixture.request_task.request
-
-    scenario.tasks.add(request)
-
-    return cast(RestApiUser, user), scenario
+class NotAnAuth:
+    pass
 
 
-def test_refresh_token_client(restapi_user: RestApiScenarioFixture, mocker: MockerFixture) -> None:
-    [user, scenario] = restapi_user
+def test_refresh_token_client(grizzly_fixture: GrizzlyFixture, mocker: MockerFixture) -> None:
+    _, user, _ = grizzly_fixture(user_type=RestApiUser)
+    assert isinstance(user, RestApiUser)
+
     decorator = refresh_token(DummyAuth)
     get_token_mock = mocker.spy(DummyAuth, 'get_token')
 
@@ -59,8 +46,8 @@ def test_refresh_token_client(restapi_user: RestApiScenarioFixture, mocker: Mock
     try:
         auth_client_context = auth_context['client']
         request_task = RequestTask(RequestMethod.POST, name='test-request', endpoint='/api/test')
-        scenario.tasks.clear()
-        scenario.tasks.add(request_task)
+        user._scenario.tasks.clear()
+        user._scenario.tasks.add(request_task)
 
         # no authentication for api, request will be called which raises NotRefreshed
         assert auth_context['provider'] is None
@@ -107,9 +94,18 @@ def test_refresh_token_client(restapi_user: RestApiScenarioFixture, mocker: Mock
         pass
 
 
-def test_refresh_token_user(restapi_user: RestApiScenarioFixture, mocker: MockerFixture) -> None:
-    [user, scenario] = restapi_user
-    decorator = refresh_token(DummyAuth)
+def test_refresh_token_user(grizzly_fixture: GrizzlyFixture, mocker: MockerFixture) -> None:
+    _, user, _ = grizzly_fixture(user_type=RestApiUser)
+    assert isinstance(user, RestApiUser)
+
+    with pytest.raises(AssertionError) as ae:
+        refresh_token('tests.unit.test_grizzly.auth.test___init__.NotAnAuth')
+    assert str(ae.value) == 'tests.unit.test_grizzly.auth.test___init__.NotAnAuth is not a subclass of grizzly.auth.RefreshToken'
+
+    # use string instead of class
+    refresh_token('AAD')
+
+    decorator = refresh_token('tests.unit.test_grizzly.auth.test___init__.DummyAuth')
     get_token_mock = mocker.spy(DummyAuth, 'get_token')
 
     auth_context = user.context()['auth']
@@ -127,8 +123,8 @@ def test_refresh_token_user(restapi_user: RestApiScenarioFixture, mocker: Mocker
     try:
         auth_user_context = auth_context['user']
         request_task = RequestTask(RequestMethod.POST, name='test-request', endpoint='/api/test')
-        scenario.tasks.clear()
-        scenario.tasks.add(request_task)
+        user._scenario.tasks.clear()
+        user._scenario.tasks.add(request_task)
 
         # no authentication for api, request will be called which raises NotRefreshed
         assert auth_user_context['username'] is None
